+++ examples/basic-mpa/src/pages/Profile/index.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Profile = () => {
+  return (
+    <>
+      <h2>Profile Page</h2>
+    </>
+  );
+};
+
+export default Profile;