+++ packages/plugin-ice-ssr/CHANGELOG.md
@@ -9,6 +9,10 @@
 
 - [feat] support setting getInitialProps with router configuration
 
+# 2.1.3
+
+- [fix] make `appConfig.app.rootId` works when ssr
+
 ## 2.1.2
 
 - [fix] not generate server bundle in Windows